+++ Configuration Management/Controls/InfobaseListView.xaml.cs
@@ -1,0 +1,11 @@
+using System.Windows.Controls;
+
+namespace Configuration_Management.Controls;
+
+public partial class InfobaseListView : UserControl
+{
+    public InfobaseListView()
+    {
+        InitializeComponent();
+    }
+}